refactor: 결과 없음 페이지 분리
